refactor(tests): fold EncapsulatedElement into Selector.resolve
Platform locators now pick android/ios inside Selector; page objects keep using Matchers and resolve().
